fix: sJson zero-init, handlers_scan -Wconversion, test arena size
- sJson.c: zero-initialize internal fields in json_make_object
  (pairs, sorted_idx, len, cap, is_sorted) and json_make_array
  (items, len, cap) to prevent use of uninitialized memory
  when arena memory is reused after destroy+create
- handlers_scan.c: add (uint32_t) cast on sizeof(req_type) in
  AOB_HANDLER macro to fix -Wconversion error
- test_sjson.c: increase arena sizes from 4096 to 65536 (64 KiB)
  and remove broken calloc workaround in test_empty_structures

+++ src/core/daemon/handlers_scan.c
@@ -112,7 +112,7 @@ memdbg_status_t handle_scan_pointer(int fd, const memdbg_packet_header_t *req,
     const req_type *aob_req = (const req_type *)body;                          \
     uint32_t pat_len = aob_req->pattern_length;                                \
     if (pat_len == 0U || pat_len > 256U) return MEMDBG_ERR_PARAM;              \
-    uint32_t expected = sizeof(req_type) + pat_len + pat_len;                  \
+    uint32_t expected = (uint32_t)sizeof(req_type) + pat_len + pat_len;                  \
     if (body_len < expected) return MEMDBG_ERR_PROTOCOL;                       \
     const uint8_t *pattern = (const uint8_t *)body + sizeof(req_type);         \
     const uint8_t *mask    = pattern + pat_len;                                \
